manifest_extraction_toolkit/src/requirements.txt
pypdf
pandas
openpyxl
pytesseract
Pillow
PyMuPDF